Update ResultId, UserId, and ChallengeId properties to use long data type in Result.cs Update EntityFramework packages and add DotNetEnv and Npgsql.EntityFrameworkCore.PostgreSQL packages in Entities.csproj Add AppDbContext class in AppDbContext.cs

+++ Entities/Models/Result.cs
@@ -9,11 +9,11 @@ namespace Challengify.Models;
 public class Result
 {
     [Key]
-    public int ResultId { get; set; }
+    public long ResultId { get; set; }
     [Required]
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     [Required]
-    public int ChallengeId { get; set; }
+    public long ChallengeId { get; set; }
     [Required, MaxLength(128)]
     public required string Name { get; set; }
     [MaxLength(512)]
